requirements/dev.txt
alembic==0.8.10
amqp==2.3.2
apache-airflow==1.10.1
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.4
bleach==2.1.4
celery==4.1.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
configparser==3.5.0
croniter==0.3.26
cryptography==2.4.2
defusedxml==0.5.0
dill==0.2.8.2
docutils==0.14
Flask==0.12.4
Flask-Admin==1.4.1
Flask-AppBuilder==1.12.2
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
flower==0.9.2
funcsigs==1.0.0
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
hmsclient==0.1.1
html5lib==1.0.1
idna==2.7
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
JayDeBeApi==1.1.1
Jinja2==2.8.1
JPype1==0.6.3
kombu==4.2.1
lockfile==0.12.2
lxml==4.2.5
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
monotonic==1.5

# temporary disabled in order to avoid to install so dependencies
#mysqlclient==1.3.13

ndg-httpsclient==0.5.1
numpy==1.15.4
ordereddict==1.1
pandas==0.23.4
paramiko==2.4.2
pendulum==1.4.4
psutil==4.4.2
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyHive==0.6.1
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
pysftp==0.2.9
python-daemon==2.1.2
python-dateutil==2.7.5
python-editor==1.0.3
python-nvd3==0.15.0
python-slugify==1.2.6
python3-openid==3.1.0
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
rdflib==4.2.2
redis==2.10.6
requests==2.20.1
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.5
SPARQLWrapper==1.8.2
SQLAlchemy==1.1.18
sshtunnel==0.1.4
tabulate==0.8.2
tagme==0.1.3
tenacity==4.8.0
thrift==0.11.0
tornado==5.1.1
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.24.1
vine==1.1.4
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.3.0
